Major changes (branch-final)
Fixed Dataseed
Added information to Dataseed

Removed DummyData

Removed all relations to DummyData

Added relations to SnurrtumlareDbContext

NOTE:  All Entities of SnurrtumlareDbContext should now be named "DbContext" for simplicity

Moved logic for searchbar to ProductsService instead of Controller.

Removed all notations of "Customer" and changed it into User. There was alot of mixups in the code that made it unclear if we were talking about a customer or adminUser.

Made changes to the properties in several models to simplify database relations and also for display simplicity.

Removed all warnings and errors by refactoring the code

minor fixes regarding typos and indents.