Added a delete function to cart
It is now possible to delete the whole cart. A popup asks for confirmation when the button is pressed. Also the ContainsProducts bool is set to false when this happens.

fix:
bool (ContainsProducts) is set to false when ever the cart is empty. This was not the case before.

change:
cleaned up the code a little to make it more readable.