Final
The user is now checked and if needed, added to the SnurrtumlareDB. If the user is already in the DB but missing an Email that will be updated to the one logged in at the moment. If nothing is missing the user is just passed on to the OrderConfirmation without any action.

If the user has all data in their profile the confirmation button will be available but if one or more fields are missing information moving on with the order confirmation will not be possible.

Functionallity to update user information will be added in another branch and commit!